+++ src/main/java/com/bicosteve/api_gateway/models/Bet.java
@@ -16,7 +16,7 @@ import java.util.List;
 @Builder
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class Bet{
-    private Integer         betId;
+    private Long         betId;
     private Long            profileId;
     private BigDecimal      stake;
     private Integer         isBonus;
